Move contents of `testing` module in `files` and `timing` modules
