+++ tum_esm_utils/__init__.py
@@ -14,7 +14,6 @@ from . import (
     processes,
     shell,
     system,
-    testing,
     text,
     timing,
     validators,
