+++ tum_esm_utils/timing.py
@@ -1,9 +1,9 @@
 """Functions used for timing or time calculations.
 
 Implements: `date_range`, `ensure_section_duration`, `set_alarm`,
-`clear_alarm`"""
+`clear_alarm`, `wait_for_condition`"""
 
-from typing import Any, Generator, List, Optional
+from typing import Any, Callable, Generator, List, Optional
 import contextlib
 import datetime
 import re
@@ -127,3 +127,24 @@ def parse_timezone_string(
     assert td is not None, f'Zone "{zone_string}" requires a datetime object to calculate the offset.'
     offset += td.total_seconds() / 3600
     return offset
+
+
+def wait_for_condition(
+    is_successful: Callable[[], bool],
+    timeout_message: str,
+    timeout_seconds: float = 5,
+    check_interval_seconds: float = 0.25,
+) -> None:
+    """Wait for the given condition to be true, or raise a TimeoutError
+    if the condition is not met within the given timeout.
+
+    `check_interval_seconds` controls, how long to wait inbetween
+    `is_successful()` calls."""
+
+    start_time = time.time()
+    while True:
+        if is_successful():
+            break
+        if (time.time() - start_time) > timeout_seconds:
+            raise TimeoutError(timeout_message)
+        time.sleep(check_interval_seconds)
